cors/credentials-flag.htm

<!DOCTYPE html>
<title>CORS - Access-Control-Allow-Credentials</title>
<meta name=author title="Odin Hørthe Omdal" href="mailto:odiho@opera.com">

<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=support.js></script>

<h1>CORS - Access-Control-Allow-Credentials</h1>
<div id=log></div>
<script>

var url = CROSSDOMAIN + 'resources/cors-cookie.php?ident='


/*
 * widthCredentials
 */
// XXX Do some https tests here as well
test(function () {
    var id = new Date().getTime(),
        client = new XMLHttpRequest()
    client.open("GET", url + id, false)
    client.send(null)
    assert_equals(client.response, "NO_COOKIE");

    client.open("GET", url + id, false)
    client.send(null)
    assert_equals(client.response, "NO_COOKIE")
}, "Don't send cookie by default");

test(function () {
    var id = new Date().getTime(),
        client = new XMLHttpRequest()

    client.open("GET", url + id, false)
    client.withCredentials = true
    client.send(null)
    assert_equals(client.response, "NO_COOKIE");

    /* We have cookie, but the browser shouldn't send */
    client.open("GET", url + id, false)
    client.withCredentials = false
    client.send(null)
    assert_equals(client.response, "NO_COOKIE")

    /* Reads and deletes the cookie */
    client.open("GET", url + id, false)
    client.withCredentials = true
    client.send(null)
    assert_equals(client.response, "COOKIE")
}, "Don't send cookie part 2");

test(function () {
    var id = new Date().getTime(),
        client = new XMLHttpRequest()

    /* Shouldn't set the response cookie */
    client.open("GET", url + id, false)
    client.withCredentials = false
    client.send(null)
    assert_equals(client.response, "NO_COOKIE");

    /* Sets the cookie */
    client.open("GET", url + id, false)
    client.withCredentials = true
    client.send(null)
    assert_equals(client.response, "NO_COOKIE")

    /* Reads and deletes the cookie */
    client.open("GET", url + id, false)
    client.withCredentials = true
    client.send(null)
    assert_equals(client.response, "COOKIE")
}, "Don't obey Set-Cookie when withCredentials=false");

function test_response_header(allow) {
    test(function () {
        var client = new XMLHttpRequest()
        client.open('GET',
            CROSSDOMAIN + 'resources/cors-makeheader.php?credentials=' + allow,
            false)
        client.withCredentials = true;
        assert_throws(null, function() { client.send() }, 'send')
    }, 'Access-Control-Allow-Credentials: ' + allow + ' should be disallowed (sync)')

    var resp_test = async_test('Access-Control-Allow-Credentials: ' + allow + ' should be disallowed (async)')
    resp_test.step(function() {
        var client = new XMLHttpRequest()
        client.open('GET',
            CROSSDOMAIN + 'resources/cors-makeheader.php?credentials=' + allow,
            true)
        client.withCredentials = true;
        client.onload = resp_test.step_func(function() {
            assert_unreached("onload")
        })
        client.onerror = resp_test.step_func(function () {
            assert_equals(client.readyState, client.DONE, 'readyState')
            resp_test.done()
        })
        client.send()
    })
}

test_response_header('TRUE')
test_response_header('True')
test_response_header('"true"')
test_response_header('false')
test_response_header('1')
test_response_header('0')

</script>
